bed_mesh: Support manipulation of meshes
Adds an extra SET parameter option to BED_MESH_PROFILE
to specify the mesh as a parameter (string form of the
dict describing the mesh).
In addition status printer.bed_mesh.profiles is added
to expose the dict of profiles defined in the module
currently in the same format as SET (after conversion
to a string) can accept.

Main goal of these changes are to allow manipulation
of the meshes from macros, to also allow storing the
meshes without needing a reset, ex. stored in a
saved variable.

Signed-off-by: Kurt Haenen <kurt.haenen@gmail.com>